Исправленный спринт 11
