+++ ImageFeed/Models/ProfileModels.swift
@@ -1,0 +1,15 @@
+//
+//  ProfileServiceModel.swift
+//  ImageFeed
+//
+//  Created by Timofey Bulokhov on 02.03.2024.
+//
+
+import Foundation
+
+struct Profile {
+    let username: String
+    let name: String
+    let loginName: String
+    let bio: String?
+}
