+++ ImageFeed/Network/Errors.swift
@@ -1,0 +1,8 @@
+//
+//  Errors.swift
+//  ImageFeed
+//
+//  Created by Timofey Bulokhov on 02.03.2024.
+//
+
+import Foundation
